First draft of minimal-mt
