+++ std/assembly/rt/index-incremental.ts
@@ -1,2 +1,3 @@
-import "rt/tlsf";
+import "rt/tlsf-base";
+import "rt/tlsf-st";
 import "rt/itcms";
